+++ src/services/settings-manager.ts
@@ -3,13 +3,13 @@ import fs from 'fs-extra';
 import config from '../config.json';
 
 export function getSettings() {
-  const settings: SettingsState = fs.existsSync(config.settingsPath)
+  const settings: Settings = fs.existsSync(config.settingsPath)
     ? fs.readJSONSync(config.settingsPath)
     : null;
 
   return settings;
 }
 
-export function setSettings(settings: SettingsState) {
+export function setSettings(settings: Settings) {
   fs.writeJsonSync(config.settingsPath, settings);
 }
